Starting to implement room. App is working

+++ app/src/main/java/com/example/falconfituser/data/api/places/PlaceRaw.kt
@@ -13,6 +13,6 @@ data class PlaceListRaw(
 data class PlaceRawAttributes(
     val title: String,
     val description: String,
-    val latitud: Float,
-    val longitud: Float,
+    val latitud: Double,
+    val longitud: Double,
 )
